Move duplicate tooltip API stories into Examples
