+++ storybook/stories/Examples/Tooltip.stories.tsx
@@ -135,3 +135,71 @@ export const SeparateDataSetsForChart = {
     );
   },
 };
+
+export const TriggerTooltipByClick = {
+  render: (tooltipArgs: Record<string, any>) => {
+    return (
+      <ResponsiveContainer width="100%" height={400}>
+        <ComposedChart data={pageData}>
+          <Tooltip {...tooltipArgs} />
+          <XAxis dataKey="name" />
+          <YAxis />
+          <Line dataKey="uv" />
+        </ComposedChart>
+      </ResponsiveContainer>
+    );
+  },
+  args: {
+    trigger: 'click',
+  },
+};
+
+interface CustomTooltipProps {
+  active?: boolean;
+  payload?: payloadType[];
+  label?: number;
+}
+
+type payloadType = {
+  value: string | number;
+  name: string;
+};
+
+const CustomContent = ({ active, payload }: CustomTooltipProps) => {
+  if (active && payload && payload.length > 0) {
+    return (
+      <div
+        style={{
+          backgroundColor: '#5b63ffe7',
+          padding: '10px',
+          borderRadius: '10px',
+          boxShadow: '1px 2px 10px -2px #7873ffb1',
+        }}
+      >
+        {payload.map((pld: payloadType) => (
+          <p
+            key={pld.name}
+            style={{
+              borderStyle: 'solid 1px',
+              fontSize: '13px',
+              fontWeight: '600',
+              fontFamily: 'sans-serif',
+              color: '#fff',
+            }}
+          >
+            {`${pld.name} : ${pld.value}`}
+          </p>
+        ))}
+      </div>
+    );
+  }
+  return null;
+};
+
+export const CustomContentExample = {
+  ...TriggerTooltipByClick,
+  args: {
+    content: <CustomContent />,
+    trigger: 'hover',
+  },
+};
